Some OpenMP parallelization of `for` loops.  Changed indices from `int` to `size_t` to avoid integer overflow with large grids.  Precompute various values.  Narrowed rapidity range to avoid NaNs from single-precision cosh and sinh; need to investigate this further, as it has been observed to affect results with larger SIGMA.  NOTE: This commit breaks compatibility with older JETSCAPE FreestreamMilneWrapper, so that freestream-milne can be built independently with its own OpenMP compiler flag (previously FreestreamMilneWrapper did `#include "FreestreamMilne.cpp"`)--so be sure to update JETSCAPE!

+++ src/Parameter.h
@@ -1,4 +1,7 @@
 #pragma once
+
+#include <stddef.h>  // for size_t
+
 struct parameters
 {
   int OUTPUTFORMAT;
@@ -9,11 +12,11 @@ struct parameters
   float ETA_FLAT;
   float SIGMA;
   float SIGMA_B;
-  int DIM_X;
-  int DIM_Y;
-  int DIM_ETA;
-  int DIM_RAP;
-  int DIM_PHIP;
+  size_t DIM_X;
+  size_t DIM_Y;
+  size_t DIM_ETA;
+  size_t DIM_RAP;
+  size_t DIM_PHIP;
   float DX;
   float DY;
   float DETA;
@@ -28,6 +31,6 @@ struct parameters
   float TAU_R;
   float ALPHA;
   //these are computed based on the chosen parameters above; they are constrained
-  int DIM;
+  size_t DIM;
   float TAU;
 };
